# Handle missing `hyps` in the exported model